+++ cmd/bootstrap/runner.go
@@ -130,9 +130,13 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		return microerror.Mask(err)
 	}
 
-	err = r.installOperators(ctx, helmClient)
-	if err != nil {
-		return microerror.Mask(err)
+	if r.flag.InstallOperators {
+		err = r.installOperators(ctx, helmClient)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	} else {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "skipping installing operators")
 	}
 
 	err = r.installAppCatalogs(ctx, k8sClients)
@@ -151,6 +155,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 func (r *runner) ensureCRDs(ctx context.Context, k8sClients k8sclient.Interface) error {
 	// Ensure Application group CRDs are created.
 	crds := []string{
+		"AppCatalogEntry",
 		"AppCatalog",
 		"App",
 		"Chart",