+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add install operators flag.
+- Create AppCatalogEntry CRD.
+
 ## [0.2.0] - 2020-10-01
 
 ### Added