+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Enable persistence in chartmuseum app CR.
+- Enable persistence in chartmuseum via its app CR.
 
 ## [0.3.0] - 2020-10-06