+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ensure PodSecurityPolicy for chartmuseum.
+- Enable ServiceAccount creation for chartmuseum.
 
 ## [0.3.1] - 2020-10-08