+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ensure PodSecurityPolicy for chartmuseum.
+- Enable ServiceAccount creation for chartmuseum.
+
 ## [0.3.1] - 2020-10-08
 
 ### Added