+++ CHANGELOG.md
@@ -16,12 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable API of chartmuseum.
 - Container build through CircleCI.
 - NetworkPolicy to enable communication with chartmuseum.
-- Patching dnsConfig and dnsPolicy of chart-operator-unique deployment.
 
 ### Fixed
 
 - Use apiextensions v3 and include replace for Giant Swarm CAPI fork. 
 - Add replace for moby v20.10.0-beta1 to fix build issue on darwin.
+- Optimize apps wait interval as app-operator has a status webhook.
 
 ## [0.3.1] - 2020-10-08