+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2020-10-30
+
 ### Added
 
 - Ensure PodSecurityPolicy for chartmuseum.
@@ -51,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add initial version that bootstraps app-operator, chart-operator and chartmuseum.
 
-[Unreleased]: https://github.com/giantswarm/apptestctl/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/giantswarm/apptestctl/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/giantswarm/apptestctl/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/giantswarm/apptestctl/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/giantswarm/apptestctl/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/apptestctl/compare/v0.1.0...v0.2.0